adding policy and fixing return value

+++ app/Http/Controllers/Api/CommentController.php
@@ -25,6 +25,10 @@ class CommentController extends Controller
 
     public function story(StoreCommentRequest $request, Story $story): JsonResponse
     {
+        if (auth()->user()->cant('create', $story)) {
+            return $this->error('Unauthorized', 403);
+        }
+
         $validated = $request->validated();
 
         try {
@@ -38,6 +42,10 @@ class CommentController extends Controller
 
     public function episode(StoreCommentRequest $request, Episode $episode): JsonResponse
     {
+        if (auth()->user()->cant('create', $episode)) {
+            return $this->error('Unauthorized', 403);
+        }
+
         $validated = $request->validated();
 
         try {
@@ -51,6 +59,10 @@ class CommentController extends Controller
 
     public function reply(StoreCommentRequest $request, Comment $comment): JsonResponse
     {
+        if (auth()->user()->cant('create', $comment)) {
+            return $this->error('Unauthorized', 403);
+        }
+        
         $validated = $request->validated();
 
         try {
@@ -64,6 +76,10 @@ class CommentController extends Controller
 
     public function update(UpdateCommentRequest $request, Comment $comment): JsonResponse
     {
+        if (auth()->user()->cant('update', $comment)) {
+            return $this->error('Unauthorized', 403);
+        }
+        
         $validated = $request->validated();
 
         try {
@@ -77,6 +93,10 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment): JsonResponse
     {
+        if (auth()->user()->cant('delete', $comment)) {
+            return $this->error('Unauthorized', 403);
+        }
+        
         try {
             $this->service->deleteComment($comment);
         } catch (\Exception $e) {
